fix: drop RDTS chain-split recovery; 31.1:14 → 31.1:15
The BIP-110 chain this machinery was written for has hard-forked to Blake2b
proof-of-work, so a flavor switch across it is no longer possible in either
direction. reconsiderblock cannot bridge a proof-of-work change: to this
binary those blocks are invalid PoW, not invalid by policy.

Removes forkRecovery.ts, the chain-recovery oneshot, the reconsiderInvalidTips
and rdtsEnforcedLastRun store keys, the Chain Recovery notification strings,
and the docs walking users through a switch that can no longer happen.
ChainTip moves to utils.ts, where the getchaintips probe that still uses it
lives.

canMigrateFrom is unchanged at <=31.1:15 — Core carries no cross-flavor edges and
never accepted a Knots flavor.
