+++ AGENTS.md
@@ -29,7 +29,7 @@ verified, tried, and decided belongs in the commit message and the PR body.
 - **Multi-branch package.** Each upstream major line lives on its own branch, checked out as a git worktree under the parent directory — `git worktree list` enumerates them. Consider every maintained worktree for any change, not just the one you are in. Release notes may legitimately differ per branch; structural changes should not.
 - **Package id is `bitcoind`, not `bitcoin-core`.** The repo and directory are named after the flavor, but `effects` calls, dependents, and `start-cli` all take `bitcoind`.
 - **`startos/utils.ts`, `startos/manifest`, and the action ids are a public API.** Fifteen sibling packages import host ids and ports from `utils` (`rpcHostId`, `rpcPort`, `peerLocalHostId`, `peerPortLocal`, `zmqHostId`, the zmq ports, `rpccookiefile`), and nine drive `autoconfig` by id. Renaming or moving one breaks their builds or their tasks — grep both registries before you do.
-- **`startos/forkRecovery.ts` is shared verbatim with every other bitcoind flavor's repo; only the callers differ.** A change here needs the same change in the Knots repos. For the same reason, never drop a key from `store.json`'s shape because this flavor does not act on it: all flavors share one store, and dropping the declaration would discard another flavor's pending state on a switch.
-- **`startos/i2pdLogFilter.ts` and `test/i2pdLogFilter.test.ts` are shared verbatim with the Knots repos**, like `forkRecovery.ts` — edit all three trees together. `npm test` runs the suite and is chained into `npm run check`, so `make` and CI gate on it. The drop list is keyed to the pinned i2pd image's exact message wording; read UPDATING.md's i2pd section before bumping that image.
+- **Never drop a key from `store.json`'s shape because this flavor does not act on it.** All bitcoind flavors share one store, and dropping the declaration would discard another flavor's pending state on a switch.
+- **`startos/i2pdLogFilter.ts` and `test/i2pdLogFilter.test.ts` are shared verbatim with the Knots repos** — edit all three trees together. `npm test` runs the suite and is chained into `npm run check`, so `make` and CI gate on it. The drop list is keyed to the pinned i2pd image's exact message wording; read UPDATING.md's i2pd section before bumping that image.
 - **`fullConfigSpec` has three hand-maintained halves and no exhaustiveness check.** A new setting must be added to `shape`, `fileToForm`, and `formToFile` in `startos/fileModels/bitcoin.conf.ts` — TypeScript will not tell you one is missing. Then select it in one of the four config actions' `filter({…})`; a field no action selects is unreachable from the UI.
 - **`setInterfaces` must not drop RPC and peer along with the conf.** Only the ZMQ and I2P exports read `bitcoin.conf`; an early return on a null read de-exports every interface, and the fifteen dependents that watch this host see them all vanish at once. A torn write is a real way to read null (Start9Labs/start-technologies#3668).
